// src/enums/pageEnum.ts
export enum PageEnum {
  // Log in
  BASE_LOGIN = '/login',
  BASE_LOGIN_NAME = 'Login',
  //redirect
  REDIRECT = '/redirect',
  REDIRECT_NAME = 'Redirect',
  // front page
  BASE_HOME = '/dashboard',
  //The home page jumps to the default route
  BASE_HOME_REDIRECT = '/dashboard/console',
  // mistake
  ERROR_PAGE_NAME = 'ErrorPage',
}